refactor: ErrorCode Enum 위치 이동
