+++ src/main/java/com/seniorway/seniorway/enums/error/ErrorCode.java
@@ -1,4 +1,4 @@
-package com.seniorway.seniorway.enums.user;
+package com.seniorway.seniorway.enums.error;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
